requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic-settings
pydantic[email]
python-multipart
gotrue
postgrest
httpx
bcrypt
python-jose[cryptography]
